+++ cmd/checker/config.go
@@ -5,8 +5,8 @@ import (
 
 	"github.com/gosexy/to"
 	"github.com/moira-alert/moira/checker"
-	"github.com/moira-alert/moira/checker/remote"
 	"github.com/moira-alert/moira/cmd"
+	"github.com/moira-alert/moira/remote"
 )
 
 type config struct {
